refactor(Controller): 분기 처리를 Controller 다형성으로 처리하도록 수정
