+++ web-application-server-gradle/src/main/java/controller/AbstractController.java
@@ -1,9 +1,14 @@
 package controller;
 
+import java.io.IOException;
 import webserver.http.MyHttpRequest;
+import webserver.http.MyHttpResponse;
 import webserver.http.Response;
 
 abstract class AbstractController implements Controller {
+    public void service(MyHttpRequest myHttpRequest, MyHttpResponse myHttpResponse) throws IOException {
+    }
+
     Response doGet(MyHttpRequest myHttpRequest) {
         return null;
     }
